added currentQuestion to components
