+++ src/components/index.js
@@ -1,3 +1,4 @@
 export { NavBar } from "./navbar/NavBar";
 export { Footer } from "./footer/Footer";
 export { Categories } from "./categories/Categories";
+export { CurrentQuestion } from "./current-question/CurrentQuestion";
